Consolidate Examples: memtable.rs and parquet_multiple_files.rs
